added layerregion.hpp and many other stuff which does not work for now

+++ src/libslic3r/Support/SupportParameters.hpp
@@ -12,7 +12,7 @@
 namespace Slic3r {
 
 class PrintObject;
-enum InfillPattern : uint8_t;
+enum InfillPattern : int;
 
 namespace FFFSupport {
